moved UserDefaults extension returned failure inside completion every time the data was nil
